Changed update metadata script to use keypair so that signMessage is available.
Refactored to remove duplicate code
