+++ packages/app/package.json
@@ -7,10 +7,10 @@
     "@heroicons/react": "^2.0.13",
     "@metaplex-foundation/js": "^0.17.12",
     "@solana/spl-token": "^0.3.6",
-    "@solana/wallet-adapter-base": "^0.9.18",
-    "@solana/wallet-adapter-react": "^0.15.23",
-    "@solana/wallet-adapter-react-ui": "^0.9.21",
-    "@solana/wallet-adapter-wallets": "^0.19.5",
+    "@solana/wallet-adapter-base": "^0.9.20",
+    "@solana/wallet-adapter-react": "^0.15.26",
+    "@solana/wallet-adapter-react-ui": "^0.9.24",
+    "@solana/wallet-adapter-wallets": "^0.19.9",
     "@solana/web3.js": "^1.66.2",
     "@sunrisestake/marinade-ts-sdk": "4.0.4-alpha.18",
     "@tailwindcss/forms": "^0.5.3",
